Update box_largest_contour.cu
it creates a box around the largest contour present in the given image. the contour can be detected using a canny, adaptive, or manual threshold functions depending on the requirement of the problem and the nature of the image. here a clean background is needed to crop out the object of interest from the image. the value of thresh can be varied for optimum results.